Canvis al MENU

+++ src/prog2/model/GeneradorVapor.java
@@ -29,7 +29,7 @@ public class GeneradorVapor {
         if (activa == false) {
             return 25;
         } else {
-            return input * 0.9;
+            return (float) (input * 0.9);
         }
     }
 }
